Add if statement for y = 0
In order to ensure that the result of the multiplication fits in an integer value
we use an if statement `else if (x <= Integer.MAX_VALUE/y)`
If input y = 0 the program would crash even though the multiplication with y=0 is possible.

I added an if statement `else if (y == 0)` that returns 0 as result.

I also added the test case for y = 0.